[backend/eth] fix a race condition
From a high level perspective cooperative settling works by
* setting up a watcher for (future) channel events and
* checking for past channel events.

The code in question did not wait for the watcher to be ready before it
began checking for past channel events. Thus, with proper timing the
code missed events if
* the watcher goroutine was started and immediately put to sleep,
* old channels events were checked,
* a channel event was emitted, and
* the watcher began waiting for events.

`TestSettler_CancelledContext()` may fail now because apparently the
simulated ETH back-end does not check the `Context` instance passed to
it.