[backend/eth] remove a test
Originally, the test was supposed to ensure that the `context.Context`
instance passed to `Settler.Settle()` is respected. The preceeding
commits cause the `Context` instance to be passed directly to the
Ethereum back-end. For testing, the simulated ethereum back-end is used
and this back-end ignores the `Context`. After discussion with Marius
van der Wijden, we decided to remove the test.