:sparkles: [eth/channel] Optimize concludeFinal for secondary requests.
In the old implementation, if two participants called Settle on their
channel at the same time, one participant's transaction would revert,
which wastes gas.

With this change, if the AdjudicatorReq is marked as 'Secondary', an
optimized protocol runs where the secondary concluder gives the other
party some blocks (currently set to 2) time to send their transaction,
saving gas for one reverted transaction in the optimistic case.

Signed-off-by: Sebastian Stammler <seb@perun.network>